Split decode module into sub-modules per type.
